Replaced rescue clause with respond_to? checks.
Exception handling is an expensive operation universally. Backtrace generation is particularly bad on JRuby. Here we guaranteed one exception for every object in the logger hierarchy. Prior to this change, this method was regularly a hotspot when profiling in JRuby.  Now it doesn't even show up in the top 50 time intensive methods.